Ionic project using ExpressAppPractice as server
This is a simple IONIC favourites list APP. The app was made from
scratch and I used HTTP post, HTTP get and HTTP delete successfully.
When using APP make sure ExpressAppPractice and mongodb is running. Make
sure that the app is running a localhost:8100. This is the default for
Ionic serve.